+++ src/core/watchQueryOptions.ts
@@ -106,7 +106,7 @@ export interface QueryOptions<TVariables = OperationVariables, TData = any> {
   /**
    * Whether to canonize cache results before returning them. Canonization
    * takes some extra time, but it speeds up future deep equality comparisons.
-   * Defaults to true.
+   * Defaults to false.
    */
   canonizeResults?: boolean;
 }